requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
reportlab
boto3
cairosvg
pdfrw
